+++ src/main/java/io/tarantool/driver/core/proxy/DeleteProxyOperation.java
@@ -48,8 +48,9 @@ public final class DeleteProxyOperation<T> extends AbstractProxyOperation<T> {
         }
 
         public DeleteProxyOperation<T> build() {
-            CRUDBaseOptions requestOptions = new CRUDBaseOptions.Builder()
+            CRUDBucketIdOptions requestOptions = new CRUDBucketIdOptions.Builder()
                     .withTimeout(options.getTimeout())
+                    .withBucketId(options.getBucketId())
                     .build();
 
             List<?> arguments = Arrays.asList(spaceName, indexQuery.getKeyValues(), requestOptions.asMap());