+++ src/test/java/io/tarantool/driver/integration/proxy/options/ProxySpaceDeleteOptionsIT.java
@@ -5,16 +5,22 @@ import io.tarantool.driver.api.TarantoolClientConfig;
 import io.tarantool.driver.api.TarantoolResult;
 import io.tarantool.driver.api.conditions.Conditions;
 import io.tarantool.driver.api.space.TarantoolSpaceOperations;
+import io.tarantool.driver.api.space.options.InsertOptions;
+import io.tarantool.driver.api.space.options.proxy.ProxyInsertOptions;
+import io.tarantool.driver.api.tuple.DefaultTarantoolTupleFactory;
 import io.tarantool.driver.api.tuple.TarantoolTuple;
+import io.tarantool.driver.api.tuple.TarantoolTupleFactory;
 import io.tarantool.driver.auth.SimpleTarantoolCredentials;
 import io.tarantool.driver.core.ClusterTarantoolTupleClient;
 import io.tarantool.driver.core.ProxyTarantoolTupleClient;
 import io.tarantool.driver.api.space.options.proxy.ProxyDeleteOptions;
 import io.tarantool.driver.integration.SharedCartridgeContainer;
+import io.tarantool.driver.mappers.DefaultMessagePackMapperFactory;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
@@ -27,6 +33,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class ProxySpaceDeleteOptionsIT extends SharedCartridgeContainer {
 
     private static TarantoolClient<TarantoolTuple, TarantoolResult<TarantoolTuple>> client;
+    private static final DefaultMessagePackMapperFactory mapperFactory = DefaultMessagePackMapperFactory.getInstance();
+    private static final TarantoolTupleFactory tupleFactory =
+            new DefaultTarantoolTupleFactory(mapperFactory.defaultComplexTypesMapper());
 
     public static String USER_NAME;
     public static String PASSWORD;
@@ -79,10 +88,35 @@ public class ProxySpaceDeleteOptionsIT extends SharedCartridgeContainer {
 
         // with option timeout
         profileSpace.delete(
-            conditions,
-            ProxyDeleteOptions.create().withTimeout(customRequestTimeout)
+                conditions,
+                ProxyDeleteOptions.create().withTimeout(customRequestTimeout)
         ).get();
         crudDeleteOpts = client.eval("return crud_delete_opts").get();
         assertEquals(customRequestTimeout, ((HashMap) crudDeleteOpts.get(0)).get("timeout"));
+    }
+
+    @Test
+    public void withBucketIdTest() throws ExecutionException, InterruptedException {
+        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> profileSpace =
+                client.space(TEST_SPACE_NAME);
+
+        TarantoolTuple tarantoolTuple = tupleFactory.create(1, null, "FIO", 50, 100);
+        Conditions condition = Conditions.equals(PK_FIELD_NAME, 1);
+
+        // with bucket id
+        Integer otherStorageBucketId = client.callForSingleResult(
+                "get_other_storage_bucket_id",
+                Collections.singletonList(tarantoolTuple.getInteger(0)),
+                Integer.class).get();
+        InsertOptions insertOptions = ProxyInsertOptions.create().withBucketId(otherStorageBucketId);
+        TarantoolResult<TarantoolTuple> insertResult = profileSpace.insert(tarantoolTuple, insertOptions).get();
+        assertEquals(1, insertResult.size());
+
+        TarantoolResult<TarantoolTuple> selectResult = profileSpace.delete(condition).get();
+        assertEquals(0, selectResult.size());
+
+        ProxyDeleteOptions deleteOptions = ProxyDeleteOptions.create().withBucketId(otherStorageBucketId);
+        selectResult = profileSpace.delete(condition, deleteOptions).get();
+        assertEquals(1, selectResult.size());
     }
 }