+++ src/test/java/io/tarantool/driver/integration/proxy/options/ProxySpaceReplaceOptionsIT.java
@@ -3,7 +3,10 @@ package io.tarantool.driver.integration.proxy.options;
 import io.tarantool.driver.api.TarantoolClient;
 import io.tarantool.driver.api.TarantoolClientConfig;
 import io.tarantool.driver.api.TarantoolResult;
+import io.tarantool.driver.api.conditions.Conditions;
 import io.tarantool.driver.api.space.TarantoolSpaceOperations;
+import io.tarantool.driver.api.space.options.ReplaceOptions;
+import io.tarantool.driver.api.space.options.proxy.ProxyDeleteOptions;
 import io.tarantool.driver.api.tuple.DefaultTarantoolTupleFactory;
 import io.tarantool.driver.api.tuple.TarantoolTuple;
 import io.tarantool.driver.api.tuple.TarantoolTupleFactory;
@@ -17,6 +20,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
@@ -37,6 +41,7 @@ public class ProxySpaceReplaceOptionsIT extends SharedCartridgeContainer {
     public static String PASSWORD;
 
     private static final String TEST_SPACE_NAME = "test__profile";
+    private static final String PK_FIELD_NAME = "profile_id";
 
     @BeforeAll
     public static void setUp() throws Exception {
@@ -89,5 +94,37 @@ public class ProxySpaceReplaceOptionsIT extends SharedCartridgeContainer {
         ).get();
         crudReplaceOpts = client.eval("return crud_replace_opts").get();
         assertEquals(customRequestTimeout, ((HashMap) crudReplaceOpts.get(0)).get("timeout"));
+    }
+
+    @Test
+    public void withBucketIdTest() throws ExecutionException, InterruptedException {
+        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> profileSpace =
+                client.space(TEST_SPACE_NAME);
+
+        TarantoolTuple tarantoolTuple = tupleFactory.create(1, null, "FIO", 50, 100);
+        Conditions condition = Conditions.equals(PK_FIELD_NAME, 1);
+
+        // without bucket id
+        TarantoolResult<TarantoolTuple> replaceResult = profileSpace.replace(tarantoolTuple).get();
+        assertEquals(1, replaceResult.size());
+
+        TarantoolResult<TarantoolTuple> deleteResult = profileSpace.delete(condition).get();
+        assertEquals(1, deleteResult.size());
+
+        // with bucket id
+        Integer otherStorageBucketId = client.callForSingleResult(
+                "get_other_storage_bucket_id",
+                Collections.singletonList(tarantoolTuple.getInteger(0)),
+                Integer.class).get();
+        ReplaceOptions replaceOptions = ProxyReplaceOptions.create().withBucketId(otherStorageBucketId);
+        replaceResult = profileSpace.replace(tarantoolTuple, replaceOptions).get();
+        assertEquals(1, replaceResult.size());
+
+        deleteResult = profileSpace.delete(condition).get();
+        assertEquals(0, deleteResult.size());
+
+        ProxyDeleteOptions deleteOptions = ProxyDeleteOptions.create().withBucketId(otherStorageBucketId);
+        deleteResult = profileSpace.delete(condition, deleteOptions).get();
+        assertEquals(1, deleteResult.size());
     }
 }